+++ src/Mod/Part/App/FeatureExtrusion.cpp
@@ -55,9 +55,9 @@ PROPERTY_SOURCE(Part::Extrusion, Part::Feature)
 Extrusion::Extrusion()
 {
     ADD_PROPERTY(Base,(0));
-    ADD_PROPERTY(Dir,(Base::Vector3f(0.0f,0.0f,1.0f)));
+    ADD_PROPERTY(Dir,(Base::Vector3d(0.0f,0.0f,1.0f)));
     ADD_PROPERTY(Solid,(false));
-    ADD_PROPERTY(TaperAngle,(0.0f));
+    ADD_PROPERTY(TaperAngle,(0.0));
 }
 
 short Extrusion::mustExecute() const
@@ -79,7 +79,7 @@ App::DocumentObjectExecReturn *Extrusion::execute(void)
         return new App::DocumentObjectExecReturn("Linked object is not a Part object");
     Part::Feature *base = static_cast<Part::Feature*>(Base.getValue());
 
-    Base::Vector3f v = Dir.getValue();
+    Base::Vector3d v = Dir.getValue();
     gp_Vec vec(v.x,v.y,v.z);
     float taperAngle = TaperAngle.getValue();
     bool makeSolid = Solid.getValue();
